♻️ 컨트롤러 이용 API 코드 리팩토링
